+++ src/main/java/org/openlmis/referencedata/dto/BasicFacilityDto.java
@@ -36,23 +36,26 @@ import java.util.Set;
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = false)
-public class MinimalFacilityDto extends BaseDto implements Facility.Exporter {
+public class BasicFacilityDto extends BaseDto implements Facility.Exporter {
 
+  private String code;
   private String name;
+  private Boolean active;
+  private Boolean enabled;
+  private FacilityTypeDto type;
+  private GeographicZoneSimpleDto geographicZone;
 
   @Override
   public void setGeographicZone(GeographicZone geographicZone) {
-    // unsupported operation
+    this.geographicZone = new GeographicZoneSimpleDto();
+    geographicZone.export(this.geographicZone);
   }
 
   @Override
   public void setType(FacilityType type) {
-    // unsupported operation
-  }
+    this.type = new FacilityTypeDto();
+    type.export(this.type);
 
-  @Override
-  public void setCode(String code) {
-    // unsupported operation
   }
 
   @Override
@@ -62,11 +65,6 @@ public class MinimalFacilityDto extends BaseDto implements Facility.Exporter {
 
   @Override
   public void setOperator(FacilityOperator operator) {
-    // unsupported operation
-  }
-
-  @Override
-  public void setActive(Boolean active) {
     // unsupported operation
   }
 
@@ -82,11 +80,6 @@ public class MinimalFacilityDto extends BaseDto implements Facility.Exporter {
 
   @Override
   public void setComment(String comment) {
-    // unsupported operation
-  }
-
-  @Override
-  public void setEnabled(Boolean enabled) {
     // unsupported operation
   }